Create README.md
Create a Java program (with multiple Java classes) to help store the information gathered on the walrus herd.

Walrus
-name:String
-age:int
-weight:double
+Walrus():void
+Walrus(String, int double):void
+getName():String
+getAge():int
+getWeight():double
+setName(String):void
+setAge(int):void
+setWeight(double):void
+compareName(String):boolean
+compareTo(Walrus):int
+toString():String

Herd
-herd:ArrayList<Walrus>
+Herd():void
+addWalrus(String, int, double):void
+addWalrus(Walrus):void
+averageAge():int
+averageWeight():int
+getWalrus(int):Walrus
+getSize():int
+findWalrus(String):Walrus
+largestWalrus():Walrus
+oldestWalrus():Walrus
+removeWalrus(String):void
+smallestWalrus():Walrus
+toString():String
+youngestWalrus():Walrus
WalrusHerdDriver

+main(String[]):void



Create a class (Walrus from the corresponding UML diagram) that handles names of walrus, the age of the walrus, and the weight of the walrus. In addition to the standard accessor and mutator methods from the instance fields, the class will have two different compare methods and a method that will output the data in a user friendly format.
Create a class (Herd from the corresponding UML diagram) that maintains the herd of walrus. It should be able to add a new walrus to the herd and remove a walrus identified by name from the herd. It should also be able to determine the average weight, the average age, the largest walrus, the smallest walrus, the youngest walrus and the oldest walrus in the herd.
Finally, create a class (WalrusHerdDriver from the corresponding UML diagram) that will allow the user to input data on say 5 walruses (create a new Walrus object based on the Walrus class for each of the walruses to inputted based on the data table below) and outputs the following information to the screen: the name and weight of the largest walrus, the name and weight of the smallest walrus, the name and age of the youngest walrus, and the name and age of the oldest walrus, average weight of the walruses and average age of the walruses.